[v1.0.1]  Fix Grid and bug with .yml files
- Fix bug where sometimes, more than one .yml file was created for the same song.
- Add buttons to increase and decrease the Grid Offset and the Grid BPM.
- Grid BPM now accepts two decimal digits.

+++ src/Assets/Scripts/WallDanceVR/YamlDeserializer.cs
@@ -39,7 +39,7 @@ public class DeserializedSong
 	public string Genere     { get; set; }
     public float AudioTime { get; set; }
     public bool Grid { get; set; }
-    public int GridBpm { get; set; }
+    public float GridBpm { get; set; }
     public float GridOffset { get; set; }
